♻️(ingestion) casse le cycle d'import celery_app → container
Extrait le bootstrap DI (get_container/init_container) dans
infrastructure/di/bootstrap.py pour que celery_app.py soit un
module feuille sans dépendance vers container.py.

container.py peut désormais importer save_raw_offer_webhook en
tête de fichier et supprimer l'import différé.
